MinorChanges

+++ FFMPEG.cs
@@ -22,36 +22,38 @@ namespace Media_Downloader
             {
                 Thread.Sleep(1000);
             }
+
             String[] rutasOrigen = File.ReadAllText(path).Split('\n');
 
-            Parallel.ForEach(rutasOrigen, new ParallelOptions { MaxDegreeOfParallelism = 10 }, s => {
-                if (!String.IsNullOrWhiteSpace(s))
-                {
-                    String input = s;
-                    String output = s.Substring(0, s.Length - s.Split('.').Last().Count()) + extensionSeleccionada;
-                    if (!File.Exists(s))
-                    {
-                        input = s.Substring(0, s.Length - s.Split('.').Last().Count()) + "mkv";
-                    }
+            Parallel.ForEach(rutasOrigen, new ParallelOptions { MaxDegreeOfParallelism = 4 }, (s, status) =>
+             {
+                 if (!String.IsNullOrWhiteSpace(s))
+                 {
+                     String input = s;
+                     String output = s.Substring(0, s.Length - s.Split('.').Last().Count()) + extensionSeleccionada;
+                     if (!File.Exists(s))
+                     {
+                         input = s.Substring(0, s.Length - s.Split('.').Last().Count()) + "mkv";
+                     }
 
-                    Process pr = new Process();
-                    pr.StartInfo = new ProcessStartInfo
-                    {
-                        FileName = FFMPEGFile,
-                        RedirectStandardOutput = true,
-                        UseShellExecute = false,
-                        CreateNoWindow = true
-                    };
+                     Process pr = new Process();
+                     pr.StartInfo = new ProcessStartInfo
+                     {
+                         FileName = FFMPEGFile,
+                         RedirectStandardOutput = true,
+                         UseShellExecute = false,
+                         CreateNoWindow = true
+                     };
                     //4audio ffmpeg.exe -y -i 'INPUT' -f FORMATO -vn -c:v copy 'OUTPUT' 
                     //4video .\ffmpeg.exe -y -i 'INPUT' -map -c:v copy 'OUTPUT'
 
                     pr.StartInfo.Arguments = "-y -i \"" + input + "\" -c:v copy \"" + output + "\"";
-                    pr.Start();
-                    pr.WaitForExit();
-                    pr.Dispose();
-                    File.Delete(input);
-                }
-            });
+                     pr.Start();
+                     pr.WaitForExit();
+                     pr.Dispose();
+                     if (!input.Equals(output)) File.Delete(input);
+                 }
+             });
             File.Delete(path);
         }
     }
